framework: transport: fix cmdu fragmentation
Current state in the transport code for fragmented packets is to send
each fragment as is, so the first fragment will not include end of
message TLV. Found out by functional testing with containers and
sniffing using wireshark, this is not in par with the ieee1905
specification.

The actual reason that end of message TLV must exist for each CMDU, is
due to the simple fact that there is no message length in the CMDU
header, hence it is impossible for a receiver to know when the fragment
ends without end of message TLV.

This commit fixes this by initially allocating and zeroing out the last
3 bytes of each fragment, thus enforcing end of message TLV for each
fragment.

In addition, add a workaround not to drop fragment if there is a
mismatch with the fragment length (more bytes after end of message TLV).
Currently the last fragment has this issue, so make that a warning for
now.
